ci: do not "cargo test" on MSRV
The doctest requires chrono-tz in
dev-dependencies. Thus, when building with "cargo
test", even if the doctest itself is not run,
chrono-tz and its dependencies are pulled in.
Unfortunately, however, cargo from our MSRV (1.48)
cannot parse the manifest of some of these. So,
the build step of "cargo test" breaks (even before
getting to run the test or ignoring it).

We still have the "cargo check" test in CI for our
MSRV.